+++ katana/payload.py
@@ -61,7 +61,9 @@ FIELD_MAPPINGS = {
     'path_delimiter': 'd',
     'deferred_calls': 'dc',
     'deprecated': 'D',
+    'duration': 'D',
     'allow_empty': 'e',
+    'end_time': 'e',
     'entity_path': 'e',
     'errors': 'e',
     'entity': 'E',
@@ -129,14 +131,15 @@ FIELD_MAPPINGS = {
     'result': 'r',
     'rollback': 'r',
     'remote_calls': 'rc',
+    'return': 'rv',
     'response': 'R',
-    'return': 'R',
     'schema': 's',
     'schemes': 's',
     'scope': 's',
     'service': 's',
     'shared': 's',
     'size': 's',
+    'start_time': 's',
     'status': 's',
     'swap': 's',
     'system': 's',
@@ -404,13 +407,18 @@ class ResponsePayload(Payload):
     name = 'response'
 
     @classmethod
-    def new(cls, version=None, status=None, body=None, headers=None):
+    def new(cls, version=None, status=None, body=None, **kwargs):
         payload = cls()
         payload.set('version', version or '1.1')
         payload.set('status', status or '200 OK')
         payload.set('body', body or '')
+
+        headers = kwargs.get('headers')
         if headers:
             payload.set('headers', headers)
+
+        if 'return_value' in kwargs:
+            payload.set('return', kwargs['return_value'])
 
         return payload